+++ src/dcspy/utils.py
@@ -255,12 +255,14 @@ def check_github_repo(git_ref: str, repo_dir: Path, repo: str, update: bool = Tr
 
     :param git_ref: Any Git reference as a string
     :param repo_dir: Local directory for a repository
-    :param repo: GitHub repository user/name
+    :param repo: GitHub repository address
     :param update: Perform update process
     :param progress: Progress callback
     """
     bios_repo = _checkout_repo(git_ref=git_ref, repo_dir=repo_dir, repo=repo, progress=progress)
-    if update:
+    is_detached = bios_repo.head.is_detached
+    LOG.debug(f'Repo at: {git_ref} with HEAD detached: {is_detached}')
+    if update and not is_detached:
         f_info = bios_repo.remotes[0].pull(progress=progress)
         LOG.debug(f'Pulled: {f_info[0].name} as: {f_info[0].commit}')
     git_ref =  git_ref.split('/')[1] if 'origin/' in git_ref else git_ref
@@ -274,7 +276,7 @@ def _checkout_repo(git_ref: str, repo_dir: Path, repo: str, progress: git.Remote
 
     :param git_ref: Any Git reference as a string
     :param repo_dir: Local repository directory
-    :param repo: Repository name
+    :param repo: Repository address
     :param progress: Progress callback
     :return: Repo object of the repository
     """
@@ -284,15 +286,34 @@ def _checkout_repo(git_ref: str, repo_dir: Path, repo: str, progress: git.Remote
     if is_git_repo(str(repo_dir)):
         bios_repo = git.Repo(repo_dir)
         all_refs = get_all_git_refs(repo_dir=repo_dir)
-        local_branch = git_ref.removeprefix('origin/')
-        if local_branch in all_refs:
-            bios_repo.git.checkout(local_branch, force=True)
+        named_git_obj = git_ref.removeprefix('origin/')
+        if named_git_obj in all_refs:
+            bios_repo.git.checkout(named_git_obj, force=True)
+        elif is_git_sha(repo=bios_repo, ref=named_git_obj):
+            bios_repo.git.checkout('--detach', named_git_obj, force=True)
         else:
-            bios_repo.git.checkout(git_ref, b=local_branch, force=True)
+            bios_repo.git.checkout(git_ref, b=named_git_obj, force=True)
     else:
         rmtree(path=repo_dir, ignore_errors=True)
-        bios_repo = git.Repo.clone_from(url=f'https://github.com/{repo}.git', to_path=repo_dir, progress=progress)  # type: ignore[arg-type]
+        bios_repo = git.Repo.clone_from(url=repo, to_path=repo_dir, progress=progress)  # type: ignore[arg-type]
     return bios_repo
+
+
+def is_git_sha(repo: git.Repo, ref: str) -> bool:
+    """
+    Check if a ref is a git commit SHA.
+
+    :param repo: Git Repository object
+    :param ref: Git commit SHA as string
+    :return: True if ref is SHA of git commit, False otherwise
+    """
+    import gitdb  # type: ignore[import-untyped]
+
+    try:
+        _  = repo.commit(ref).hexsha
+        return True
+    except gitdb.exc.BadName:
+        return False
 
 
 def check_dcs_bios_entry(lua_dst_data: str, lua_dst_path: Path, temp_dir: Path) -> str:
@@ -480,30 +501,33 @@ def _fetch_system_info(conf_dict: dict[str, Any]) -> str:
     dcs = check_dcs_ver(dcs_path=Path(str(conf_dict['dcs'])))
     bios_ver = check_bios_ver(bios_path=str(conf_dict['dcsbios']))
     repo_dir = Path(str(conf_dict['dcsbios'])).parents[1] / 'dcs-bios'
-    git_ver, head_commit = _fetch_git_data(repo_dir=repo_dir)
+    git_ver, head_commit, remote_url = _fetch_git_data(repo_dir=repo_dir)
     lgs_dir = '\n'.join([
         str(Path(dir_path) / filename)
         for dir_path, _, filenames in walk('C:\\Program Files\\Logitech Gaming Software\\SDK')
         for filename in filenames
     ])
-    return f'{__version__=}\n{name=}\n{pyver=}\n{pyexec=}\n{dcs=}\n{bios_ver=}\n{git_ver=}\n{head_commit=}\n{lgs_dir}\ncfg={pformat(conf_dict)}'
+    return f'{__version__=}\n{name=}\n{pyver=}\n{pyexec=}\n{dcs=}\n{bios_ver=}\n{remote_url=}\n{git_ver=}\n{head_commit=}\n{lgs_dir}\ncfg={pformat(conf_dict)}'
 
 
-def _fetch_git_data(repo_dir: Path) -> tuple[Sequence[int], str]:
+def _fetch_git_data(repo_dir: Path) -> tuple[Sequence[int], str, str]:
     """
     Fetch the Git version and SHA of HEAD commit.
 
     :param repo_dir: Local directory for repository
-    :return: Tuple of (a version) and SHA of HEAD commit
+    :return: Tuple of (a version), SHA of HEAD commit and remote URL
     """
     try:
         import git
         git_ver = git.cmd.Git().version_info
-        head_commit = str(git.Repo(repo_dir).head.commit)
+        repo = git.Repo(repo_dir)
+        head_commit = str(repo.head.commit)
+        remote_url = repo.remotes.origin.url
     except (git.exc.NoSuchPathError, git.exc.InvalidGitRepositoryError, ImportError):
         git_ver = (0, 0, 0, 0)
         head_commit = 'N/A'
-    return git_ver, head_commit
+        remote_url = 'N/A'
+    return git_ver, head_commit, remote_url
 
 
 def _get_dcs_log(conf_dict: dict[str, Any]) -> Path: